// src/app/Models/Pet.ts
export class Pet {
    constructor(){
        
    }
    petId: number | any;
    petName: string | any;
    species: string | any;
    dateOfBirth: string | any;
    gender: string | any;
    age: string | any;
    bloodGroup: string | any;
    allergies: string | any;
    petOwnerId: number | any;
    parentName: string | any;
    image:string | any;
}